rename Error component
